+++ lib/active_support/cache/leveled_cache_store.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ActiveSupport
+  module Cache
+    LeveledCacheStore = LeveledCache::Store
+  end
+end