+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Support ActiveSupport::Cache.lookup\_store
+
 ## [0.1.0] - 2024-02-13
 
 - Initial release